fix: 🚑로그인 하지 않으면 글 작성불가. (#151)

+++ src/post/UpdatePost.tsx
@@ -92,13 +92,15 @@ const UpdatePost = () => {
   };
 
   const handleDeletePost = () => {
-    mutate({
-      url: `${END_POINT}/posts/delete`,
-      method: 'delete',
-      data: {
-        id: postId,
-      },
-    }).then(() => navigate(`/channel/${channelId}`, { replace: true }));
+    if (confirm('글을 삭제하시겠습니까?')) {
+      mutate({
+        url: `${END_POINT}/posts/delete`,
+        method: 'delete',
+        data: {
+          id: postId,
+        },
+      }).then(() => navigate(`/channel/${channelId}`, { replace: true }));
+    }
   };
 
   const handleOnClickUploadImage = async (e: React.ChangeEvent<HTMLInputElement>) => {
